Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
qrcode[pil]==7.4.2
pydantic==2.5.0
passlib[bcrypt]==1.7.4
alembic==1.13.0